Add query for test boot from volume snapshot bug 1489581
Since this is using test_ids I'm not sure how many hits it is,
it's quite a bit when you filter things in logstash.o.o but
the delete message here shows up in successful runs so I can't
just use that.  I'm using test_ids since it's really only
failing on this new test that was recently added.

Change-Id: I88804485f86452c3f357c8fa8650a294ed968ecb
Related-Bug: #1489581