Add note in query for bug 1249065
Add a note to the query for bug 1249065 to remind
us not to remove that query until kilo-eol happens.

The final fix 73a68abb28b9ad5f4e9dee2781cdc5203bcb5937 in
nova contains an RPC API version change so we aren't going
to backport it to stable/kilo, so this bug can still show up
there infrequently.

Change-Id: Ic08a5e12a1376a4ea8c441a40f447b21d269c8d9
Related-Bug: #1249065