Add a query for libvirt error when 'TestBootVolumePattern' fails
Not an alarming number of hits (17 hits in n-cpu log in the last seven
days), but would be useful to track down the root cause.

There are also related errors from iSCSI daemon from the 'syslog' of the
said job ('gate-grenade-dsvm-multinode')

"session 2 login rejected: Initiator failed authentication with target"

which has about 152 occurrences in the last seven days.

Related-Bug: #1522488

Change-Id: Id160292199fba195b28dfad7150dece38fb3a50e