Add query for manila-neutron job timeout bug 1528597
90 hits in 7 days, check and gate, master and stable/liberty,
all failures. The job timeout is set to 70 minutes but the test
timeout is set to 500 seconds, and there is at least one test
taking over 400 seconds, and this mostly hits on the OVH nodes,
which are the slowest that infra uses.

So slow tests plus slow nodes equals timeout. But the manila team
should investigate what's so slow about the tests anyway.

Change-Id: I52ccee0141024f1182bbccfb740016527d21f876
Related-Bug: #1528597