Add query for nova py34 fake image service bug 1531988
Not pretty but this is a common failure with nova py34 jobs
because of mox being used in the stubbed out fake image service
in a bunch of tests.

178 hits in 7 days, check and gate, all failures.

There are patches up to skip some known racy tests and a patch
is up to remove the mox usage in the fake image service stub.

Change-Id: I0ed003bd60624cabe47ae2d9db7204ac7f32a6dc
Related-Bug: #1531988