Add query for pip 8 release explosion
This commit adds the query for the fallout from the pip 8 release.
The fixes/caps have already landed for this but this is just to get
it off the uncategorized page. Luckily the message format change from
when it was deprecated so we can uniquely identify when pip8 was used
on a distutils package vs when <8 was used.

Change-Id: Iacfb3d7837d0376e8993c01f791f810de3ebf0be