Add query for Ironic bug 1537754
We've been seeing an error in Ironic's tempest-BaremetalBasicOps test
where q-svc reports DB Error RouterNotFound, causing this test to fail.

Of note, the same error is intentionally created by a Neutron tempest
job, so I have added this test name to the query to limit it to
situations where it should not be failing.

Change-Id: I3b7664e9ca6a96e2f70278b7c2e3acbc0f9ef12e