Add allow-nonvoting key
By default we add 'voting:1' to all queries so they are filtered
from the graph.

There are cases where we want to track bugs for non-voting jobs
in the graph, so add an 'allow-nonvoting' key for special bugs
that we want to track while we're trying to stabilize a job.

Updates the docs, adds a test query and unit test for the query
loader, fixes a bad query that was setting voting:1 already, and
removes the TODO in the query for bug 1539271.

Change-Id: I61b1d4b0faaa3f20fc5b04aa3392c68c61990866