Mark non-voting jobs in the graph
Commit e456a7afca0168a1cc408a3c30d6067b8c2ec169 added the
'allow-nonvoting' key to query yaml files so that we can
show non-voting job failures in the graph.

This change builds on that by displaying in the graph,
for non-voting job queries only based on that key, when a
job is non-voting with a simple "Voting: False" line.

Since the default behavior is to filter out non-voting
job results in all of the queries, we don't show anything
special in the graph for voting jobs since it would just
clutter up the output (non-voting is the exceptional case
we want to display).

Change-Id: Ibd75c6244abd10ad7cc491b4453339ad326a11ed