Add query for nova unshelve race bug 1611008
5 hits in 7 days, check and gate, master branch only
right now. This is a race in Tempest tests that use
a shared server resource where it looks like a shelve/unshelve
is setting the instance to ACTIVE/none state before the
instance.host field is updated in the DB which makes the
next test to run on that shared server fail (because the
compute RPC doesn't know where to cast the message).

Change-Id: Iaae74a7b7533c58b5a7fdb1dd30814e6f66a0170
Related-Bug: #1611008