Add grenade to the integrated gate projects list
This commit add grenade to the list of projects in the integrated gate
list. While it technically doesn't have the job tag set, this is only
because it doesn't regular tempest jobs. It's still a part of the
integrated gate for the same reasons devstack and tempest are so we
should include it in the list.

Change-Id: If8ee91d4c9c8e519b045617fe2e3c7a09fad231d