Add query for d-g ping check failures.
This adds a query for bug 1694501 which will track when devstack-gate
ping checks to the mirror host fail. This fails jobs because we fail
early if we cannot talk to the mirror host in expectation that the rest
of the job will fail in a less clear manner due to networking problems.

Change-Id: I16783d942a7b51c527eac6f80f0a8daa933e0352