Add query for bug 1701088
This commit adds a query for the gate issue where glance uploads fail
with a connection reset in uwsgi. This has 245 hits in the last week
all failures in both gate and check, except for a single success with
the hit and a traceback in glance. But the corresponding test didn't
fail. (likely a bug in the test, or a negative test)

Change-Id: I32c1c40339bdae342a706647efd7f0cb9f6c9bb6
Related-Bug: #1701088