Use clearer query for 1326813
This bug, 1326813, is an old one that was about rackspace and hpcloud.
We don't use hpcloud anymore (it doesn't exist) and yet we were getting
hits on this query. Digging in the bug is actually an error of
easy_install (distutils) talking to its pypi mirror. I have updated the
bug to better reflect this and this updates the query to query on
distutils specific errors to keep them tightly coupled together.

This should make it much easier to understand what is going on when we
see hits for this bug.

Change-Id: Id1c5211702f9de889a87facba9694c7a8c52e1e1
Related-bug: 1326813