Refactor how we track pip errors
There are two major class of pip error, the first is fatal the second is
not.

First we have the complete inability to find a distribution for a
package. This can happen due to network errors or the package simply not
existing. Track that in 1449136 our long standing pip had a problem bug.

Second there is the pip had a retriable network problem and is retrying.
By default 5 attempts are made and typically pip will eventually
succeed. If it does not succeed it logs the error from the first issue.
We track this retry issue in 1708686 because while not fatal it may
indicate larger underlying problems we should be aware of but also don't
want it distracting from actual failures in jobs.

Change-Id: I2dbe4b887ed868bd57c58a79a6f2b5600366e116
Related-bug: 1449136
Related-bug: 1708686