Add query for rebuild timeout cellsv1 bug 1709985
9 hits in 7 days, check queue only but multiple changes,
all failures.

This is restricted to the cells v1 job since we have
a similar bug in non-cellsv1 jobs where nova times out
waiting to get a network changed/vif plugged event from
neutron and that causes the rebuild to fail. The cellsv1
job wouldn't have that problem as it doesn't use neutron.

Change-Id: Ibc8ccc5fcf6d94a88a89565b907bbf3edd63ce47
Related-Bug: #1709985