Add query for latent cells v1 rebuild race bug 1709985
This query existed at one point and then was removed,
probably because we rename the job and it wasn't hitting
anymore as a result. This still shows up from time to
time as new rebuild tests are added to tempest so we
might as well track it.

Currently 6 hits in the check queue on multiple changes.

Change-Id: Ie8715b6c43a57a8c8521c96ea2bbd5d8266cf950