Add query for mysql opportunistic test bug 1793364
This is grasping at something that hits on these failures
since we don't have multiline on the console log failures
for unit tests, but we have an IndexError that seems to
be causing a greenthread switch and I think it might be
related to sqla returning an empty result set when the
caller is expecting results and not handling it properly.

Anyway, this gives 1890 hits in 7 days, check and gate,
master, rocky and queens branches, all failures.

Change-Id: If224014d6f353fda1258988d834cadc7e3818be6