Contrain query to prevent e-r from OOMing
The previous 1793364 query matched tripleo's error.txt file in which
they grep and store all the errors found during a job run. Unfortunately
the number and size of these results was such that the e-r processes
would OOM trying to process them.

Adjust the query to look only at the job-output.txt files to fix this.

Change-Id: I82a6a9cb6a442fe39c0ec5faa31599c7c4fc5134