Adjust query for bug 1800472
The previous query was too restrictive and was missing failures
on other tests. After triaging the bug, it looks like the real
issue is in the API code and a warning is logged which is our
clue to the failure, so the query is updated here to use that
warning message.

Change-Id: I759e81e01721a3325dcce917732524dfcb5b2e0a