Update query for bug 1721093
The original query no longer works because it assumes that unreachable
is on the same line as the rest of the ssh message, which is not true
anymore. This makes the new search much more reliable as it will clearly
not miss to spot such an error.

The old query yields 3 hits in 7 days, all failures, check and gate, but
the new query yields 1303 hits in 7 days, all failures, check and gate.

Change-Id: I7739a0481497cdc8be658e902d526564a11130a7