Add query for libvirt functional evacuate test bug 1806126
This isn't great but it's a functional test so there are
no service logs. It's a race that hits in a couple of new
tests added in change:

  https://review.openstack.org/#/c/591733/

And the best thing probably to identify it is image-backend
stuff that happens in the libvirt driver before it either
blows up or hangs and times out.

32 hits in 48 hours, check and gate, all failures.

Change-Id: Iaaa91df6020fe353c0b44a5f4f0a63fffb896339