Update query for bug 1808010
We have noticed that there are job failures which emit the warning
about not being able to add the networking route but do not have
any messages about disk space issues.

Update the query to catch the other cases so we may investigate the
root cause (slow metadata API response) further.

Related-Bug: #1808010

Change-Id: I21e98d459472d18d39aba66fd5a63f2ef7f8d5e9