Add query for cinder tooz connection error bug 1810526
68 hits in 7 days, check and gate, not all failures though,
actually only about 62% failure, but I'm not sure how to
track this and get 100% failure rates as it seems sometimes
when tooz connections to etcd drop out it doesn't affect tests
and maybe that's because cinder isn't being tested at the time,
or they are expected like during grenade runs (although I'm not
sure if we restart etcd in grenade).

Anyway, it would be good either way to track this since it should
not be happening and indicates we're overloading etcd and need
to sort that out somehow.

Change-Id: I8d49d05d753e775e249d70e0dbcb00594e295ee7